requirements.txt
# Core dependencies
mininet==2.3.0
pyyaml==6.0.1
python-dateutil==2.8.2
recurring-ical-events==2.0.2
icalendar==5.0.11

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Networking
pyroute2==0.7.12
netaddr==0.10.1
scapy==2.5.0

# Observability
prometheus-client==0.19.0
python-json-logger==2.0.7
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Database/Storage
sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1

# Task scheduling
apscheduler==3.10.4
croniter==2.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==24.1.1
pylint==3.0.3
mypy==1.8.0
types-PyYAML==6.0.12.12
types-python-dateutil==2.8.19.14

# Utilities
click==8.1.7
rich==13.7.0
requests==2.31.0
tabulate==0.9.0
jinja2==3.1.3

# Container/Docker support
docker==7.0.0

# HTTP client
httpx==0.26.0
aiofiles==23.2.1

# Cryptography (for WireGuard key generation)
cryptography==42.0.0

